Fixed warnings

+++ src/CMemoryFile.cpp
@@ -135,7 +135,7 @@ uint64_t CMemoryFile::Read(uint8_t* buffer, uint64_t size)
     uint64_t maxSize = std::min(size, bufferSize);
     if (maxSize > 0)
     {
-        memcpy(buffer, m_Data.data(), maxSize);
+        memcpy(buffer, m_Data.data(), (size_t)maxSize);
     }
     else
     {
@@ -155,9 +155,9 @@ uint64_t CMemoryFile::Write(const uint8_t* buffer, uint64_t size)
     uint64_t bufferSize = Size() - Tell();
     if (size > bufferSize)
     {
-        m_Data.resize(m_Data.size() + (size - bufferSize));
+        m_Data.resize((size_t)(m_Data.size() + (size - bufferSize)));
     }
-    memcpy(m_Data.data() + Tell(), buffer, size);
+    memcpy(m_Data.data() + Tell(), buffer, (size_t)size);
     
     return size;
 }
